+++ examples/Blinky1/Blinky1/Blinky1Sm.cpp
@@ -40,8 +40,7 @@ void Blinky1Sm_start(Blinky1Sm* self)
     // ROOT behavior
     // uml: TransitionTo(ROOT.InitialState)
     {
-        // Step 1: Exit states until we reach `ROOT` state (Least Common Ancestor for transition).
-        exit_up_to_state_handler(self, ROOT_exit);
+        // Step 1: Exit states until we reach `ROOT` state (Least Common Ancestor for transition). Already at LCA, no exiting required.
         
         // Step 2: Transition action: ``.