+++ examples/LaserTagMenu1/LaserTagMenu1/LaserTagMenu1Sm.c
@@ -163,8 +163,7 @@ void LaserTagMenu1Sm_start(LaserTagMenu1Sm* self)
     // ROOT behavior
     // uml: TransitionTo(ROOT.InitialState)
     {
-        // Step 1: Exit states until we reach `ROOT` state (Least Common Ancestor for transition).
-        exit_up_to_state_handler(self, ROOT_exit);
+        // Step 1: Exit states until we reach `ROOT` state (Least Common Ancestor for transition). Already at LCA, no exiting required.
         
         // Step 2: Transition action: ``.