+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
-## WIP
+## [0.7.12-alpha]
 ### Added
 - draw.io - convert more HTML breaks (`<br style=...>`, `</br>`) to new line characters.
 - draw.io - convert more special characters to regular space characters.